Convert BaseLayout to TS
